+++ src/script/api/qscriptengine.cpp
@@ -1010,7 +1010,7 @@ QVariant QScriptEnginePrivate::jscValueToVariant(JSC::ExecState *exec, JSC::JSVa
     QVariant v(targetType, (void *)0);
     if (convertValue(exec, value, targetType, v.data()))
         return v;
-    if (uint(targetType) == QVariant::LastType)
+    if (targetType == QVariant::LastType)
         return toVariant(exec, value);
     if (isVariant(value)) {
         v = variantValue(value);
